Fix OCI hooks without optional connections or SDK
Principal authentication does not require an Airflow connection, and missing-SDK coverage must remain active in minimal provider environments.
