+++ providers/oracle/tests/unit/oracle/hooks/test_oci_optional_dependency.py
@@ -1,0 +1,55 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+from __future__ import annotations
+
+import subprocess
+import sys
+from textwrap import dedent
+from unittest import mock
+
+import pytest
+
+from airflow.providers.common.compat.sdk import AirflowOptionalProviderFeatureException
+from airflow.providers.oracle.hooks.base_oci import _get_oci_sdk
+
+
+def test_get_oci_sdk_requires_optional_extra():
+    with mock.patch.dict(sys.modules, {"oci": None}):
+        with pytest.raises(
+            AirflowOptionalProviderFeatureException,
+            match=r"pip install 'apache-airflow-providers-oracle\[oci\]'",
+        ):
+            _get_oci_sdk()
+
+
+def test_hook_modules_import_without_optional_oci_sdk():
+    subprocess.run(
+        [
+            sys.executable,
+            "-c",
+            dedent(
+                """
+                import sys
+
+                sys.modules["oci"] = None
+                import airflow.providers.oracle.hooks.base_oci
+                import airflow.providers.oracle.hooks.generative_ai
+                """
+            ),
+        ],
+        check=True,
+    )
